% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroupR.R
\name{regroupR}
\alias{regroupR}
\title{Regroup faulty aligned peaks}
\usage{
regroupR(aligned.peaks, list.to.regroup, min.samp.grp = 1,
  max.dupli.prop = 0.1, maxClust = 10)
}
\arguments{
\item{aligned.peaks}{The aligned peaks data.}

\item{list.to.regroup}{The peak indices of groups to regroup (the groups, indicated by their peakIndex, in 1 list item will be merged and regrouped).}

\item{min.samp.grp}{The minimal amount of samples needed te form a group.}

\item{max.dupli.prop}{The maximal duplication proportion allowed for a group to be considered a sigle group.}

\item{maxClust}{The maximum number of clusters (depth of the tree).}
}
\value{
Returns a data frame with regrouped peaks.
}
\description{
If there are groups misaligned by the peakAligner function, they will be realigned by also using the peak signal to noise ratio as well as the ppm values.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}

